Split up the transaction module a little bit
